order book fix

+++ exchange_context/model/depth.go
@@ -4,7 +4,7 @@ import "sort"
 
 type Depth struct {
 	Symbol    string      `json:"s"`
-	Timestamp int         `json:"T,int"`
+	Timestamp int64       `json:"T,int"`
 	Bids      [][2]Number `json:"b"`
 	Asks      [][2]Number `json:"a"`
 }
